requirements.txt
streamlit
python-dotenv
google-generativeai
pdfplumber
PyPDF2
Pillow
pandas
